+++ weblate/trans/models/__init__.py
@@ -27,7 +27,7 @@ from weblate.trans.models.label import Label
 from weblate.trans.models.pending import PendingUnitChange
 from weblate.trans.models.project import CommitPolicyChoices, Project
 from weblate.trans.models.report import Report
-from weblate.trans.models.suggestion import Suggestion, Vote
+from weblate.trans.models.suggestion import Suggestion, SuggestionAddResult, Vote
 from weblate.trans.models.translation import Translation
 from weblate.trans.models.unit import Unit
 from weblate.trans.models.variant import Variant
@@ -57,6 +57,7 @@ __all__ = [
     "Project",
     "Report",
     "Suggestion",
+    "SuggestionAddResult",
     "Translation",
     "Unit",
     "Variant",